refactor(customers): migrate to entity toolkit pattern with business sub-router
- Convert flat customers.router.ts to customers.router/ directory
- createEntityRouter(customerServerSpec) provides entity toolkit
- Existing procedures bundled in business.router.ts (internals unchanged)
- submitCustomerAge stays at root with publicProcedure + manual token validation
- All frontend callers updated: customersRouter.xxx → customersRouter.business.xxx
