+++ src/features/lead-sources-admin/ui/components/all-customers-section.tsx
@@ -30,7 +30,7 @@ export function AllCustomersSection() {
   const { setModal, open: openModal } = useModalStore()
 
   const pagination = usePaginatedQuery<Record<string, never>, CustomerTableRow>(
-    trpc.customersRouter.list.queryOptions,
+    trpc.customersRouter.business.list.queryOptions,
     {},
     {
       paramPrefix: 'all',
@@ -41,7 +41,7 @@ export function AllCustomersSection() {
   )
 
   const updateCreatedAt = useMutation(
-    trpc.customersRouter.updateCreatedAt.mutationOptions({
+    trpc.customersRouter.business.updateCreatedAt.mutationOptions({
       onSuccess: () => {
         toast.success('Created date updated')
         invalidateCustomer()
